Reorganised conftest.py files for the best.
